updated conclusion componet needsPermit -> needpermit
